Implemented Confrim Dialog

+++ index.js
@@ -10,6 +10,8 @@ let searchForm = document.getElementById("search-form"),
 
 let tasksSection = document.querySelector("#tasks-section");
 
+let confirmDialog = document.querySelector("#confirm-dialog");
+
 let deleteIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="currentColor" class="bi bi-trash-fill" viewBox="0 0 16 16">
 <path d="M2.5 1a1 1 0 0 0-1 1v1a1 1 0 0 0 1 1H3v9a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V4h.5a1 1 0 0 0 1-1V2a1 1 0 0 0-1-1H10a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1H2.5zm3 4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 .5-.5zM8 5a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7A.5.5 0 0 1 8 5zm3 .5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 1 0z"/>
 </svg>`;
@@ -36,6 +38,18 @@ function emptyTasks(){
         taskPar.classList.add("no-tasks");
         taskPar.textContent = "No Tasks Added";
         tasksSection.insertAdjacentElement("beforeend",taskPar);
+};
+
+function showConfirmDialog(callback){
+    confirmDialog.classList.add("show-item");
+    document.querySelector("#confirm-btn").addEventListener("click",function(){
+        this.closest("#confirm-dialog").classList.remove("show-item");
+        callback(true)
+    });
+    document.querySelector("#cancel-btn").addEventListener("click",function(){
+        this.closest("#confirm-dialog").classList.remove("show-item");
+        callback(false);
+    });
 };
 
 function searchTasks(){
@@ -102,14 +116,19 @@ function mapTasks(tasks){
     
             tasksSection.insertAdjacentElement("afterbegin",taskdiv);
             document.querySelector(".delete-icon").addEventListener("click",function(){
-                this.closest(".task-body").remove();
-                let tasks = JSON.parse(localStorage.getItem("tasks")),
-                    newTasks = tasks.filter(t=>t.task != task.task);
-                localStorage.setItem("tasks",JSON.stringify(newTasks));
-                if (newTasks.length == 0){
-                    emptyTasks();
-                    localStorage.setItem("tasks",null);
-                }
+                document.querySelector("#confirm-task").textContent = task.task;
+                showConfirmDialog((res)=>{
+                    if (res){
+                        this.closest(".task-body").remove();
+                        let tasks = JSON.parse(localStorage.getItem("tasks")),
+                            newTasks = tasks.filter(t=>t.task != task.task);
+                        localStorage.setItem("tasks",JSON.stringify(newTasks));
+                        if (newTasks.length == 0){
+                            emptyTasks();
+                            localStorage.setItem("tasks",null);
+                        }
+                    }
+                });
             });
             document.querySelector(".done-icon").addEventListener("click",function(){
                 if (!taskName.classList.contains("add-task")){
